+++ src/RcppExports.cpp
@@ -5,6 +5,11 @@
 
 using namespace Rcpp;
 
+#ifdef RCPP_USE_GLOBAL_ROSTREAM
+Rcpp::Rostream<true>&  Rcpp::Rcout = Rcpp::Rcpp_cout_get();
+Rcpp::Rostream<false>& Rcpp::Rcerr = Rcpp::Rcpp_cerr_get();
+#endif
+
 // BMNLLInner
 double BMNLLInner(double a, double b, NumericVector zetas, NumericVector fi, NumericVector ti);
 RcppExport SEXP _RJafroc_BMNLLInner(SEXP aSEXP, SEXP bSEXP, SEXP zetasSEXP, SEXP fiSEXP, SEXP tiSEXP) {